+++ src/preload/index.ts
@@ -7,6 +7,7 @@ const bridge: Bridge = {
   connection: () => ipcRenderer.invoke('connection:get'), saveSecrets: value => ipcRenderer.invoke('connection:save', value), connectGmail: () => ipcRenderer.invoke('connection:connect'), disconnectGmail: () => ipcRenderer.invoke('connection:disconnect'),
   listMail: value => ipcRenderer.invoke('mail:list', value), readMail: id => ipcRenderer.invoke('mail:read', id), translateMail: id => ipcRenderer.invoke('mail:translate', id),
   testAI: () => ipcRenderer.invoke('ai:test'), digests: () => ipcRenderer.invoke('digest:list'), summarize: () => ipcRenderer.invoke('digest:run'), openExternal: url => ipcRenderer.invoke('external:open', url),
+  themes: () => ipcRenderer.invoke('themes:list'), installTheme: () => ipcRenderer.invoke('themes:install'), installExampleTheme: id => ipcRenderer.invoke('themes:install-example', id), removeTheme: id => ipcRenderer.invoke('themes:remove', id),
   onChanged: callback => { const listener = () => callback(); ipcRenderer.on('workspace:changed', listener); return () => { ipcRenderer.removeListener('workspace:changed', listener) } }
 }
 contextBridge.exposeInMainWorld('workstation', bridge)